+++ packages/tokens/js/components/dark/input.js
@@ -21,6 +21,9 @@ export default {
   minWidth: 152,
   paddingHorizontal: 16,
   paddingVertical: 12,
+  stepper: {
+    gap: 4,
+  },
   textArea: {
     height: 96,
   },