+++ packages/tokens/js/components/dark/table.d.ts
@@ -13,9 +13,6 @@ declare const _default: {
   readonly headerCell: {
     readonly borderWidth: 2;
     readonly gap: 8;
-    readonly neutral: {
-      readonly backgroundColor: '#30302c';
-    };
     readonly paddingHorizontal: 12;
     readonly paddingVertical: 16;
   };